Minor tweaks

+++ src/player.cpp
@@ -9,9 +9,9 @@ Projectile Player::shoot(const Player& player, const int mousePosX, const int mo
 
 void Player::draw() {
 	// Draw the player
-	DrawRectangle((int)x, (int)y, width, width, BLUE);
+	DrawRectangle((int)x, (int)y, (int)width, (int)width, BLUE);
 
 	//Draw the health bar
 	DrawRectangle((int)x - 300, (int)y + 350, 600, 50, BLACK);
-	DrawRectangle((int)x - 295, (int)y + 355, 590 * ((float)health/maxHealth), 40, GREEN);
+	DrawRectangle((int)x - 295, (int)y + 355, (int)(590 * ((float)health/maxHealth)), 40, GREEN);
 }
